SPOGOU is now an addon
First upload after developing an addon version of SPOGOU. Yeah, should probably have used some git tool to keep track of all changes. Live and learn.
It's still a bound addon, so future development is going to be to make it completely standalone, and of course smoother corners (figuratively).
Left the old script version of SPOGOU as a separate branch.